Refactor LdpRequest, make newResource() synchronous.
