+++ lib/data-storage/ldp-backend-fs/ldp-file-store.js
@@ -5,10 +5,6 @@ const fs = require('fs-extra')
 const { v1: uuidv1 } = require('uuid')
 const { URL } = require('url')
 const { logger } = require('../../logger')
-const { promisify } = require('util')
-const rdf = require('rdflib')
-const parseRdf = promisify(rdf.parse)
-const serializeGraph = promisify(rdf.serialize)
 const { extensions } = require('mime-types')
 const http = require('http')
 const https = require('https')
@@ -18,7 +14,7 @@ const { LdpStore } = require('../ldp-store')
 const { LdpFileResource, LdpFileContainer } = require('./ldp-file-resource')
 const { LdpTarget } = require('../ldp-target')
 
-const { DEFAULT_RDF_TYPE, DEFAULT_ENCODING } = require('../../defaults')
+const { DEFAULT_ENCODING } = require('../../defaults')
 const AVAILABLE_CHARSETS = ['utf8']
 
 class LdpFileStore extends LdpStore {
@@ -27,6 +23,8 @@ class LdpFileStore extends LdpStore {
    *
    * @param options.fs {object} Expects the `fs-extra` API
    * @param options.mapper {LegacyResourceMapper}
+   *
+   * Inherited from superclass:
    * @param options.host {SolidHost}
    */
   constructor (options) {
@@ -53,7 +51,7 @@ class LdpFileStore extends LdpStore {
     const {
       path,
       contentType
-    } = await this.mapper.mapUrlToFile({ url: target.url })
+    } = this.mapper.mapUrlToFile({ url: target.url })
 
     let exists
     let fsStats
@@ -95,6 +93,8 @@ class LdpFileStore extends LdpStore {
    * Creates a new Resource or Container instance. Used for creating new
    * resources rather than mapping to existing ones (hence no `fsStats` param).
    *
+   * Used in multi-file upload on POST.
+   *
    * TODO: Make this sync (after making mapUrlToFile sync).
    *
    * @param target
@@ -105,14 +105,14 @@ class LdpFileStore extends LdpStore {
    *
    * @returns {LdpFileResource|LdpFileContainer}
    */
-  async newResource ({ target, path, contentType, encoding, isContainer }) {
+  newResource ({ target, path, contentType, encoding, isContainer }) {
     contentType = contentType || target.bodyContentType
 
     if (!path) {
       const {
         path: mappedPath,
         contentType: mappedContentType
-      } = await this.mapper.mapUrlToFile({ url: target.url })
+      } = this.mapper.mapUrlToFile({ url: target.url })
       path = mappedPath
       contentType = contentType || mappedContentType
     }
@@ -131,9 +131,6 @@ class LdpFileStore extends LdpStore {
    * value, and is done against a container).
    * If the target resource exists, appends a UUID to the target name (so that
    * POSTs only create resources, and not overwrite).
-   *
-   * TODO: Currently this is bound to the File System. Need to decouple it and
-   * move it to superclass, LdpStore.
    *
    * @param [slug] {string} Suggested name of file or container to be created
    * @param container {LdpFileContainer} Parent container (in which target is to
@@ -374,28 +371,6 @@ class LdpFileStore extends LdpStore {
   }
 
   /**
-   * Loads and parses the container's `.meta` resource.
-   *
-   * @param container {LdpFileContainer}
-   *
-   * @returns {Promise<IndexedFormula>} Resolves with the parsed `.meta` graph.
-   */
-  async readContainerMeta ({ container }) {
-    let graph
-
-    try {
-      const metaTarget = new LdpTarget({ url: container.target.aclUrl })
-      const metaResource = await this.resource({ target: metaTarget })
-      graph = await this.loadParsedGraph({ resource: metaResource })
-    } catch (error) {
-      // do nothing, this is likely a 'not found' error
-      graph = rdf.graph() // new/empty graph
-    }
-
-    return graph
-  }
-
-  /**
    * Deletes a given resource.
    *
    * @param resource {LdpFileResource}
@@ -463,74 +438,8 @@ class LdpFileStore extends LdpStore {
     await this.fs.ensureDir(dirname(path))
     return this.fs.writeFileSync(path, blob)
   }
-
-  /**
-   * Loads an RDF file from disk, parses it, and returns the resulting graph.
-   *
-   * @param [resource] {LdpResource}
-   *
-   * @param [graph] {IndexedFormula}
-   * @param [contentType] {string}
-   *
-   * @throws {Error} File system errors (file does not exist, `EACCES` etc)
-   * @throws {Error} Parse errors
-   *
-   * @returns {Promise<IndexedFormula>} Resolves with parsed graph
-   */
-  async loadParsedGraph ({
-    resource, graph = rdf.graph(),
-    contentType = DEFAULT_RDF_TYPE
-  }) {
-    const rawGraphContents = await this.readBlob({ resource })
-    if (!rawGraphContents) {
-      return graph
-    }
-    return parseRdf(rawGraphContents, graph, resource.target.url, contentType)
-  }
-
-  /**
-   * Serializes and writes a graph to a given resource, and returns the original
-   * (non-serialized) graph.
-   *
-   * @param resource {LdpFileResource}
-   *
-   * @param graph {IndexedFormula}
-   * @param [contentType] {string}
-   *
-   * @return {Promise<IndexedFormula>}
-   */
-  async writeGraph ({ resource, graph, contentType = DEFAULT_RDF_TYPE }) {
-    // target, kb, base, contentType
-    const { url } = resource.target
-    const graphRdf = await serializeGraph(null, graph, url, contentType)
-
-    return this.writeBlob({ resource, blob: graphRdf })
-  }
-
-  /**
-   * Attempts to fetch a graph first from the local store, and then via fetch,
-   * and returns the parsed graph
-   *
-   * TODO: Only does local for now, not remote. Will use fetchRemoteGraph() in
-   *   the future
-   *
-   * @param [url] {string}
-   * @param [resource] {LdpResource}
-   * @param graph {IndexedFormula}
-   * @param contentType {string}
-   * @returns {Promise<IndexedFormula>}
-   */
-  async fetchGraph ({ url, resource, graph, contentType = DEFAULT_RDF_TYPE }) {
-    if (!resource) {
-      const target = new LdpTarget({ url })
-      resource = await this.resource({ target })
-    }
-
-    return this.loadParsedGraph({ resource, graph, contentType })
-  }
 }
 
 module.exports = {
-  LdpStore,
   LdpFileStore
 }
