+++ lib/data-storage/ldp-backend-fs/legacy-resource-mapper.js
@@ -17,7 +17,7 @@ class LegacyResourceMapper extends ResourceMapper {
   }
 
   // Maps the request for a given resource and representation format to a server file
-  async mapUrlToFile ({ url }) {
+  mapUrlToFile ({ url }) {
     return { path: this._getFullPath(url), contentType: this._getContentTypeByExtension(url) }
   }
