+++ lib/data-storage/ldp-request.js
@@ -30,15 +30,16 @@ class LdpRequest {
    * @param headerMeta {Metadata}
    */
   constructor ({ target, resource, headers, bodyStream, ldpStore, credentials = null, host, headerMeta }) {
+    this.headers = headers
+    this.headerMeta = headerMeta
+    this.bodyStream = bodyStream
+
     this.target = target
     this.resource = resource
-    this.headers = headers
-    this.bodyStream = bodyStream
     this.ldpStore = ldpStore
     // this.body = body ?
     this.credentials = credentials
     this.host = host
-    this.headerMeta = headerMeta
   }
 
   /**
@@ -578,8 +579,8 @@ class LdpPostRequest extends LdpRequest {
         const target = new LdpTarget({
           name: filename, url: fileUrl, bodyContentType: contentType
         })
-        ldpStore.newResource({ target, contentType })
-          .then(resource => ldpStore.createResource({ resource, fromStream: fileStream }))
+        const resource = ldpStore.newResource({ target, contentType })
+        return ldpStore.createResource({ resource, fromStream: fileStream })
           .catch(error => busboy.emit('error', error))
       })
       busboy.on('error', error => {
