+++ lib/data-storage/ldp-store.js
@@ -1,6 +1,11 @@
 'use strict'
 
-const { ACL_SUFFIX, META_SUFFIX } = require('../defaults')
+const { ACL_SUFFIX, META_SUFFIX, DEFAULT_RDF_TYPE } = require('../defaults')
+const { promisify } = require('util')
+const rdf = require('rdflib')
+const parseRdf = promisify(rdf.parse)
+const serializeGraph = promisify(rdf.serialize)
+const { LdpTarget } = require('./ldp-target')
 
 class LdpStore {
   /**
@@ -21,6 +26,42 @@ class LdpStore {
    */
   async resource ({ target }) {
     throw new Error('resource() must be implemented in subclass.')
+  }
+
+  /**
+   * Creates a new Resource or Container instance. Used for creating new
+   * resources rather than mapping to existing ones (hence no `fsStats` param).
+   *
+   * Used in multi-file upload on POST.
+   *
+   * @param target
+   * @param [path]
+   * @param [contentType]
+   * @param [encoding] {string}
+   * @param [isContainer] {boolean}
+   *
+   * @returns {LdpFileResource|LdpFileContainer}
+   */
+  newResource ({ target, path, contentType, encoding, isContainer }) {
+    throw new Error('newResource() must be implemented in subclass.')
+  }
+
+  /**
+   * Creates an LdpTarget instance for a POST request (which includes a `slug`
+   * value, and is done against a container).
+   * If the target resource exists, appends a UUID to the target name (so that
+   * POSTs only create resources, and not overwrite).
+   *
+   * @param [slug] {string} Suggested name of file or container to be created
+   * @param container {LdpContainer} Parent container (in which target is to
+   *   be created)
+   * @param headerMeta {Metadata} Result of parsing `Link:` header
+   * @param bodyContentType {string} Mime type of incoming request body
+   *
+   * @returns {LdpTarget}
+   */
+  async targetFromSlug ({ slug = uuidv1(), container, headerMeta, bodyContentType }) {
+    throw new Error('targetFromSlug() must be implemented in subclass.')
   }
 
   /**
@@ -57,6 +98,28 @@ class LdpStore {
    */
   async ensureContainer ({ container }) {
     throw new Error('ensureContainer() must be implemented in subclass.')
+  }
+
+  /**
+   * Loads and parses the container's `.meta` resource.
+   *
+   * @param container {LdpContainer}
+   *
+   * @returns {Promise<IndexedFormula>} Resolves with the parsed `.meta` graph.
+   */
+  async readContainerMeta ({ container }) {
+    let graph
+
+    try {
+      const metaTarget = new LdpTarget({ url: container.target.aclUrl })
+      const metaResource = await this.resource({ target: metaTarget })
+      graph = await this.loadParsedGraph({ resource: metaResource })
+    } catch (error) {
+      // do nothing, this is likely a 'not found' error
+      graph = rdf.graph() // new/empty graph
+    }
+
+    return graph
   }
 
   /**
@@ -121,7 +184,7 @@ class LdpStore {
   }
 
   /**
-   * Loads an RDF resource from the store, parses it, and returns the resulting
+   * Loads an RDF file from the store, parses it, and returns the resulting
    * graph.
    *
    * @param [resource] {LdpResource}
@@ -129,10 +192,61 @@ class LdpStore {
    * @param [graph] {IndexedFormula}
    * @param [contentType] {string}
    *
+   * @throws {Error} Storage errors (resource does not exist)
+   * @throws {Error} Graph parse errors.
+   *
    * @returns {Promise<IndexedFormula>} Resolves with parsed graph
    */
-  async loadParsedGraph ({ resource, graph, contentType }) {
-    throw new Error('loadParsedGraph() must be implemented in subclass.')
+  async loadParsedGraph ({
+    resource, graph = rdf.graph(),
+    contentType = DEFAULT_RDF_TYPE
+  }) {
+    const rawGraphContents = await this.readBlob({ resource })
+    if (!rawGraphContents) {
+      return graph
+    }
+    return parseRdf(rawGraphContents, graph, resource.target.url, contentType)
+  }
+
+  /**
+   * Serializes and writes a graph to a given resource, and returns the original
+   * (non-serialized) graph.
+   *
+   * @param resource {LdpResource}
+   *
+   * @param graph {IndexedFormula}
+   * @param [contentType] {string}
+   *
+   * @return {Promise<IndexedFormula>}
+   */
+  async writeGraph ({ resource, graph, contentType = DEFAULT_RDF_TYPE }) {
+    // target, kb, base, contentType
+    const { url } = resource.target
+    const graphRdf = await serializeGraph(null, graph, url, contentType)
+
+    return this.writeBlob({ resource, blob: graphRdf })
+  }
+
+  /**
+   * Attempts to fetch a graph first from the local store, and then via fetch,
+   * and returns the parsed graph
+   *
+   * TODO: Only does local for now, not remote. Will use fetchRemoteGraph() in
+   *   the future
+   *
+   * @param [url] {string}
+   * @param [resource] {LdpResource}
+   * @param graph {IndexedFormula}
+   * @param contentType {string}
+   * @returns {Promise<IndexedFormula>}
+   */
+  async fetchGraph ({ url, resource, graph, contentType = DEFAULT_RDF_TYPE }) {
+    if (!resource) {
+      const target = new LdpTarget({ url })
+      resource = await this.resource({ target })
+    }
+
+    return this.loadParsedGraph({ resource, graph, contentType })
   }
 
   /**
