+++ package.json
@@ -68,7 +68,7 @@
     "from2": "^2.3.0",
     "fs-extra": "^8.1.0",
     "handlebars": "^4.7.3",
-    "inquirer": "^7.0.5",
+    "inquirer": "^7.0.6",
     "li": "^1.3.0",
     "mime-types": "^2.1.26",
     "nano": "^8.2.1",
@@ -83,7 +83,7 @@
     "solid-namespace": "^0.3.0",
     "standard-http-error": "^2.0.1",
     "ulid": "^2.3.0",
-    "uuid": "^7.0.1",
+    "uuid": "^7.0.2",
     "valid-url": "^1.0.9",
     "vhost": "^3.0.2"
   },
